updated request handler to manage timeout and added related tests. moved req reply stuff to separate file. added
